Add .astype('int64') to Numpy array examples since np.random.randint(num_classes, size=num_valid_samples) gives int32 arrays instead of int64 on Linux

+++ pytoune/framework/model.py
@@ -54,12 +54,12 @@ class Model:
             # Our training dataset with 800 samples.
             num_train_samples = 800
             train_x = np.random.randn(num_train_samples, num_features).astype('float32')
-            train_y = np.random.randint(num_classes, size=num_train_samples)
+            train_y = np.random.randint(num_classes, size=num_train_samples).astype('int64')
 
             # Our validation dataset with 200 samples.
             num_valid_samples = 200
             valid_x = np.random.randn(num_valid_samples, num_features).astype('float32')
-            valid_y = np.random.randint(num_classes, size=num_valid_samples)
+            valid_y = np.random.randint(num_classes, size=num_valid_samples).astype('int64')
 
             pytorch_module = torch.nn.Linear(num_features, num_classes) # Our network
